Include a test suite for the python components
Since most of the logic is being performed in python it was easiest to
use pytest which I am familiar with to perform the testing of the logic.
This is going to pick up the majority of the bugs and forms a framework
which I can use to develop the project by specifying functionality.