Ensure all paths and filenames are python Paths
I am using the python pathlib.Path representation for all the paths
within the python application. Since I was trying to compare the
internal paths with the strings passed by vim there were some issues
with compatibility.

Additionally I have added simple support for logging to a file which
aided in the identification of this issue.